[#163] Fixes based on PR comments.
Use an abstract class for the tests
